further break up table into fields
